+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 **New Modules:**
 
-- ntnx_karbon_clusters_node_pools - Create,Update and Delete Node pools with the provided configuration.
+- ntnx_karbon_clusters_node_pools - Create,Update and Delete worker node pools with the provided configuration.
 - ntnx_ndb_tags_info - info module for ndb tags info